+++ src/components.d.ts
@@ -615,6 +615,10 @@ export namespace Components {
          */
         "size": 'small' | 'medium' | 'large';
         /**
+          * Enables spell checking on the input.
+         */
+        "spellcheck": boolean;
+        /**
           * The input's step attribute.
          */
         "step": number;
@@ -2065,6 +2069,10 @@ declare namespace LocalJSX {
           * The input's size.
          */
         "size"?: 'small' | 'medium' | 'large';
+        /**
+          * Enables spell checking on the input.
+         */
+        "spellcheck"?: boolean;
         /**
           * The input's step attribute.
          */